+++ src/settings.ts
@@ -33,6 +33,15 @@ export namespace Settings {
       };
     }
 
+    // Make export Format only onb Joplin > 2.9.12 public
+    let exportFormatPublic = false;
+    if (
+      joplinVersionInfo !== null &&
+      (await helper.versionCompare(joplinVersionInfo.version, "2.9.12")) >= 0
+    ) {
+      exportFormatPublic = true;
+    }
+
     await joplin.settings.registerSettings({
       path: pathSettings,
       backupRetention: {
@@ -187,6 +196,21 @@ export namespace Settings {
         label: "Backup plugins",
         description: "Backup plugin jpl files",
       },
+      exportFormat: {
+        value: "jex",
+        type: SettingItemType.String,
+        section: "backupSection",
+        isEnum: true,
+        public: exportFormatPublic,
+        label: "Export format",
+        advanced: true,
+        options: {
+          jex: "Jex",
+          md_frontmatter: "MD Frontmatter",
+          raw: "RAW",
+        },
+        description: "Backup format for the notes.",
+      },
       singleJexV2: {
         value: true,
         type: SettingItemType.Bool,
@@ -205,6 +229,15 @@ export namespace Settings {
         advanced: true,
         label: "Single JEX",
         description: "Old setting, for compatibility and upgrade only.",
+      },
+      execFinishCmd: {
+        value: "",
+        type: SettingItemType.String,
+        section: "backupSection",
+        public: true,
+        advanced: true,
+        label: "Command on Backup finish",
+        description: "Execute command when backup is finished.",
       },
       backupVersion: {
         value: 0,